+++ src/v2019C1/extensions/index.ts
@@ -1,0 +1,5 @@
+import * as Function from './Function'
+
+export const EXTENSIONS = {
+	Function,
+} as const