+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.6] - 2026-01-29
+
+### Added
+
+- `exportSclFile` updated: `withDownload` parameter that triggers the download of the file
+
+## [0.0.5] - 2026-01-29
+
+### Added
+
+- `exportSclFile` updated: add custom extension based on configuration `supportedExtensions`
+
+## [0.0.4] - 2026-01-29
+
+### Added
+
+- `extractTo` extensions: extract any function to a new database at the right level (ie : `Substation`, `VoltageLevel` or `Bay`)
+
 ## [0.0.3] - 2026-01-29
 
 ### Added